change: separate header file for FNL
